+++ org.eclipse.draw2d/src/org/eclipse/draw2d/text/TextFlow.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2000, 2025 IBM Corporation and others.
+ * Copyright (c) 2000, 2026 IBM Corporation and others.
  *
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License 2.0 which is available at
@@ -22,10 +22,10 @@ import org.eclipse.pde.api.tools.annotations.NoExtend;
 
 import org.eclipse.draw2d.ColorConstants;
 import org.eclipse.draw2d.Graphics;
-import org.eclipse.draw2d.TextUtilities;
 import org.eclipse.draw2d.geometry.Dimension;
 import org.eclipse.draw2d.geometry.Point;
 import org.eclipse.draw2d.geometry.Rectangle;
+import org.eclipse.draw2d.internal.DrawableTextUtilities;
 
 /**
  * An inline flow figure that renders a string of text across one or more lines.
@@ -724,21 +724,11 @@ public class TextFlow extends InlineFlow {
 	 * @return a <code>FlowUtilities</code> instance
 	 * @since 3.4
 	 */
-	@SuppressWarnings("static-method")
 	protected FlowUtilities getFlowUtilities() {
+		if (getTextUtilities() instanceof DrawableTextUtilities textUtilities) {
+			return textUtilities.getFlowUtilities();
+		}
 		return FlowUtilities.INSTANCE;
-	}
-
-	/**
-	 * Gets the <code>TextUtilities</code> instance to be used in measurement
-	 * calculations.
-	 *
-	 * @return a <code>TextUtilities</code> instance
-	 * @since 3.4
-	 */
-	@SuppressWarnings("static-method")
-	protected TextUtilities getTextUtilities() {
-		return TextUtilities.INSTANCE;
 	}
 
 }
